Header en Footer zijn nu components :D
